move test/e2e/framework/lifecycle/ test/framework/cloud/gcp
