+++ test/e2e/framework/BUILD
@@ -127,7 +127,6 @@ filegroup(
         "//test/e2e/framework/job:all-srcs",
         "//test/e2e/framework/kubectl:all-srcs",
         "//test/e2e/framework/kubelet:all-srcs",
-        "//test/e2e/framework/lifecycle:all-srcs",
         "//test/e2e/framework/log:all-srcs",
         "//test/e2e/framework/metrics:all-srcs",
         "//test/e2e/framework/network:all-srcs",
